Support editing users invitations
