+++ app/(dashboard)/users/UserMenu.tsx
@@ -45,15 +45,12 @@ export const UserMenu = ({ user }: UserMenuProps) => {
         onDelete={handleDelete}
       />
       <RowDropdownMenu>
-        {/* TODO: Support editing invitations */}
-        {user.resourceType === 'user' && (
-          <DropdownMenuItem asChild>
-            <Link href={routes.users.user(user.resourceId)}>
-              <Pencil />
-              Edit
-            </Link>
-          </DropdownMenuItem>
-        )}
+        <DropdownMenuItem asChild>
+          <Link href={routes.users.user(user.resourceId)}>
+            <Pencil />
+            Edit
+          </Link>
+        </DropdownMenuItem>
         <DropdownMenuItem
           onClick={deleteConfirm.open}
           disabled={authUser.auth.uid === user.uid}
