+++ app/(dashboard)/users/[id]/page.tsx
@@ -10,13 +10,17 @@ import { Users } from 'lucide-react'
 import { revalidatePath } from 'next/cache'
 import { notFound } from 'next/navigation'
 import { allowTypes, getAuthenticatedOnlyApp } from '@/modules/firebase/guards'
-import { mapAuthData } from '@/modules/firebase/user'
-import { getDocData, getDocsData, UserType } from '@/modules/firebase/utils'
+import {
+  getDocDataOrThrow,
+  getDocsData,
+  UserType,
+} from '@/modules/firebase/utils'
 import { routes } from '@/modules/routes'
 import { getUserName } from '@/packages/design-system/src/modules/auth/user'
 import { PageTitle } from '@/packages/design-system/src/molecules/DashboardLayout'
 import { DashboardLayout } from '../../DashboardLayout'
 import { UserForm, type UserFormSchema } from '../UserForm'
+import { getUserData } from '@/modules/user/queries'
 
 interface UserPageProps {
   params: { id: string }
@@ -24,15 +28,9 @@ interface UserPageProps {
 
 const UserPage = async ({ params }: UserPageProps) => {
   await allowTypes([UserType.admin, UserType.owner])
-  const { refs, docRefs } = await getAuthenticatedOnlyApp()
+  const { refs } = await getAuthenticatedOnlyApp()
   const userId = params.id
-  const allAuthData = await mapAuthData({ userIds: [userId] }, (data, id) => ({
-    uid: id,
-    email: data.auth.email,
-    displayName: data.auth.displayName,
-  }))
-  const authUser = allAuthData.at(0)
-  const user = await getDocData(docRefs.user(userId))
+  const { authUser, user, resourceType } = await getUserData(userId)
   if (!authUser || !user || user.type === UserType.patient) {
     notFound()
   }
@@ -41,21 +39,37 @@ const UserPage = async ({ params }: UserPageProps) => {
   const updateUser = async (form: UserFormSchema) => {
     'use server'
     const { docRefs, callables } = await getAuthenticatedOnlyApp()
-    await callables.updateUserInformation({
-      userId,
-      data: {
-        auth: {
-          displayName: form.displayName,
-          email: form.email,
-        },
-      },
-    })
-    const userRef = docRefs.user(userId)
-    await updateDoc(userRef, {
+    const authData = {
+      displayName: form.displayName,
+      email: form.email,
+    }
+    const userData = {
       invitationCode: form.invitationCode,
       organization: form.organizationId ?? deleteField(),
       type: form.type,
-    })
+    }
+    if (resourceType === 'user') {
+      await callables.updateUserInformation({
+        userId,
+        data: {
+          auth: authData,
+        },
+      })
+      await updateDoc(docRefs.user(userId), userData)
+    } else {
+      const invitation = await getDocDataOrThrow(docRefs.invitation(userId))
+      await updateDoc(docRefs.invitation(userId), {
+        auth: {
+          ...invitation.auth,
+          ...authData,
+        },
+        user: {
+          ...invitation.user,
+          ...userData,
+        },
+      })
+    }
+
     revalidatePath(routes.users.index)
   }
