+++ modules/user/queries.tsx
@@ -7,7 +7,9 @@
 //
 import { query, where } from 'firebase/firestore'
 import { getAuthenticatedOnlyApp } from '@/modules/firebase/guards'
+import { mapAuthData } from '@/modules/firebase/user'
 import {
+  getDocData,
   getDocDataOrThrow,
   getDocsData,
   type Invitation,
@@ -69,4 +71,37 @@ export const getUserOrganizationsMap = async () => {
       (organization) => [organization.id, organization] as const,
     ),
   )
+}
+
+/**
+ * Gets user or invitation data
+ * */
+export const getUserData = async (userId: string) => {
+  const { docRefs } = await getAuthenticatedOnlyApp()
+  const user = await getDocData(docRefs.user(userId))
+  if (user) {
+    const allAuthData = await mapAuthData(
+      { userIds: [userId] },
+      (data, id) => ({
+        uid: id,
+        email: data.auth.email,
+        displayName: data.auth.displayName,
+      }),
+    )
+    const authUser = allAuthData.at(0)
+    return { user, authUser, resourceType: 'user' as const }
+  }
+  const invitation = await getDocData(docRefs.invitation(userId))
+  return {
+    user: invitation?.user,
+    authUser:
+      invitation?.auth ?
+        {
+          uid: userId,
+          email: invitation.auth.email ?? null,
+          displayName: invitation.auth.displayName ?? null,
+        }
+      : undefined,
+    resourceType: 'invitation' as const,
+  }
 }
